pyproject.toml
[project]
name = "kazan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleach>=6.2.0",
    "django>=5.2.6",
    "django-markdownx>=4.0.9",
    "django-storages[s3]>=1.14.6",
    "django-treebeard>=4.7.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
